Scrum-90
This commit contains scrum-90 to scrum-95, the super administrator can create, edit and delete a municipality. Only those with the role super admin are able to make these changes and also see the "Municipality management" , when a new Municipality is created a new user is created for the admin of that Muni and in the edit page there is only the possibility to change the name, email, responsible and address.