Add end reached animation to notifications modal
Introduces a new Lottie animation for when the end of notifications is reached. Updates the notifications modal to display this animation using a new enum value and animation options. Also adds logic in the notification HTTP service to track when the end is reached and updates the notification list accordingly.